feat!: make docker image manipulation easier

+++ Dockerfile
@@ -7,7 +7,7 @@ ARG PYTHON_VERSION=3.10.0
 RUN apt-get update && apt-get install -y \
     curl locales software-properties-common git build-essential libssl-dev zlib1g-dev libbz2-dev libreadline-dev \
     libsqlite3-dev wget llvm libncursesw5-dev xz-utils tk-dev libxml2-dev \
-    libxmlsec1-dev libffi-dev ca-certificates && \
+    libxmlsec1-dev libffi-dev ca-certificates wget gpg && \
     apt-get clean && rm -rf /var/lib/apt/lists/*
 
 RUN locale-gen en_US.UTF-8 && \
@@ -23,12 +23,19 @@ RUN bash -l -c "pyenv install $PYTHON_VERSION && pyenv global $PYTHON_VERSION"
 
 RUN pip install numpy lxml
 
-RUN add-apt-repository universe && \
-    rm -rf /var/lib/apt/lists/*
+RUN add-apt-repository universe
 
 RUN curl -sSL https://raw.githubusercontent.com/ros/rosdistro/master/ros.key -o /usr/share/keyrings/ros-archive-keyring.gpg
 
+RUN wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > packages.microsoft.gpg
+
+RUN install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings/packages.microsoft.gpg
+
 RUN sh -c 'echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/ros-archive-keyring.gpg] http://packages.ros.org/ros2/ubuntu $(. /etc/os-release && echo $UBUNTU_CODENAME) main" > /etc/apt/sources.list.d/ros2.list'
+
+RUN sh -c 'echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | tee /etc/apt/sources.list.d/vscode.list > /dev/null'
+
+RUN rm -f packages.microsoft.gpg
 
 RUN apt-get update && apt-get install -y \
   python3-flake8-docstrings \
@@ -57,13 +64,17 @@ RUN apt-get update && apt-get install -y \
   evtest \
   ros-humble-twist-mux \
   ros-humble-rviz2 \
-  ros-humble-gazebo-ros2-control && \
-  rm -rf /var/lib/apt/lists/*
+  ros-humble-gazebo-ros2-control  \
+  apt-transport-https  \
+  terminator && rm -rf /var/lib/apt/lists/*
+
+RUN apt-get update && apt-get install -y code
 
 COPY . /TrailblazerML/
 
 RUN echo "source /opt/ros/humble/setup.bash" >> ~/.bashrc && \
     echo "source /usr/share/colcon_argcomplete/hook/colcon-argcomplete.bash" >> ~/.bashrc && \
+    echo "alias code='code --no-sandbox --user-data-dir /TrailblazerML/.vscode-root'" >> ~/.bashrc && \
     bash -c "source ~/.bashrc" && \
     bash -c "source ~/.bashrc && bash /TrailblazerML/scripts/dowload_rover_stl.sh"
